pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "aether"
dynamic = ["version"]
description = "Proxy server enabling Claude Code to work with OpenAI-compatible API providers"
readme = "README.md"
authors = [
    {name = "Aether", email = "noreply@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "uvicorn>=0.34.0",
    "gunicorn>=23.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.54.0",
    "httpx[socks]>=0.28.1",
    "sqlalchemy>=2.0.43",
    "alembic>=1.16.5",
    "bcrypt>=4.3.0",
    "pyjwt>=2.10.1",
    "certifi>=2025.6.15",
    "cryptography>=45.0.7",
    "psycopg2-binary>=2.9.10",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "loguru>=0.7.3",
    "tiktoken>=0.5.0",
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "aiosmtplib>=4.0.2",
    "redis>=5.0.0",
    "prometheus-client>=0.20.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[project.urls]
Homepage = "https://github.com/fawney19/Aether"
Repository = "https://github.com/fawney19/Aether.git"
Issues = "https://github.com/fawney19/Aether/issues"

[project.scripts]
aether = "src.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "factory-boy>=3.2.0",
    "faker>=18.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pyinstaller>=6.15.0",
    "hatch-vcs>=0.5.0",
]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/_version.py"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = "tools/debug"
# 忽略项目内部模块的 import-untyped 警告
ignore_missing_imports = true
# SQLAlchemy mypy 插件
plugins = ["sqlalchemy.ext.mypy.plugin"]

# SQLAlchemy 相关模块放宽类型检查（模型未使用 Mapped 注解）
[[tool.mypy.overrides]]
module = ["src.services.usage.service", "src.models.database"]
disable_error_code = ["arg-type", "assignment"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "-v"
]
norecursedirs = ["tools/debug"]
